feat: useHandleImageChange Hook 분리
